query backend set up
