+++ server/routes/camp.js
@@ -1,0 +1,10 @@
+const express = require('express');
+const camp = express.Router();
+const campController = require('../controllers/campController');
+
+camp.post('/query',
+  campController.query,
+  (req, res) => {
+    
+  }
+)
