+++ src/App.jsx
@@ -23,7 +23,6 @@ class App extends Component {
             waterFront: false,
             queriedGrounds: [],
         }
-        this.queryCampground = this.queryCampground.bind(this);
         this.petOnChange = this.petOnChange.bind(this)
         this.waterHookOnChange = this.waterHookOnChange.bind(this)
         this.sewerHookOnChange = this.sewerHookOnChange.bind(this)
@@ -31,56 +30,27 @@ class App extends Component {
         this.query = this.query.bind(this);
     }
 
-
-    
-
-    
-    queryCampground(e){
-        e.preventDefault()
-        console.log(this.state)
-        console.log('here');
-        let apiString = 'http://api.amp.active.com/camping/campgrounds?'
-        if(this.state.pet === true){
-            apiString +='&pets=3010'
-        }
-        if(this.state.sewerHook == true){
-            apiString += '&sewer=3007'
-        }
-        if(this.state.waterHook === true){
-            apiString += '&water=3007'
-        }
-        if(this.state.waterFront === true){
-            apiString += '&waterfront=3011'
-        }
-        const campOptions = {
-            url: apiString += 'APIKEYHERE',
-            method: 'GET',
+    query(){
+        fetch('/camp/query', {
+            method: 'POST',
             headers: {
-              'Accept': 'application/json',
-              'Content-Type': 'application/json;charset=UTF-8',
-            }
-          }
-          axios(campOptions)
-            .then(response => {
-              console.log(response);
-              const stringRes = response.data;
-              // console.log("this is json stringified response: ", stringRes)
-              // const tyler = parser.toJson(response);
-              // console.log('thisistyler: ', tyler);
-              // parser.parseString
-              let superParse;
-              parseString(stringRes, function (err, result) {
-                superParse = result.resultset.result;
-              });
-              console.log('hello: ', superParse);
-              fs.writeFileSync(path.resolve(__dirname, '../database/camp.json'), JSON.stringify(superParse));
-              console.log('writesuccess');
-              res.locals.wines = response.data.results;
-              return next();
+              'Content-Type': 'application/json'
+            },
+            body: JSON.stringify({
+                pet: this.state.pet,
+                waterFront: this.state.waterFront,
+                waterHook: this.state.waterHook,
+                sewerHook: this.state.sewerHook
             })
-            .catch(err => console.log('fileController.getWine: axios fetch error: ', err));
+        })
+        .then(res => res.json())
+        .then(data => {
+            const newState = Object.assign({}, this.state);
+            newState.queriedGrounds = data;
+            this.setState(newState);
+        })
     }
-
+    
     petOnChange(){
         console.log('petOnChange called')
         if (this.state.pet === false){
@@ -146,27 +116,6 @@ class App extends Component {
                 username: user,
                 password: password,
             })
-        })
-    }
-
-    query(){
-        fetch('/camp/query', {
-            method: 'POST',
-            headers: {
-              'Content-Type': 'application/json'
-            },
-            body: JSON.stringify({
-                pet: this.state.pet,
-                waterFront: this.state.waterFront,
-                waterHook: this.state.waterHook,
-                sewerHook: this.state.sewerHook
-            })
-        })
-        .then(res => res.json())
-        .then(data => {
-            const newState = Object.assign({}, this.state);
-            newState.queriedGrounds = data;
-            this.setState(newState);
         })
     }
